Fix MSVC compiler warnings
reg.: 'argument': conversion from 'size_t' to 'int', possible loss of data (C4267)

+++ src/notation/internal/notationinteraction.cpp
@@ -2293,7 +2293,7 @@ bool NotationInteraction::dragStandardElement(const PointF& pos, Qt::KeyboardMod
 
         // Where "m" is the number of page elements and "n" is the number of elements accepting a drop...
         // O(log m) operation
-        m_droppableTree.initialize(page->pageBoundingRect(), page->elements().size());
+        m_droppableTree.initialize(page->pageBoundingRect(), static_cast<int>(page->elements().size()));
 
         // O(m log n) operation
         for (EngravingItem* elem : page->elements()) {
